services/analytics/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
google-cloud-firestore==2.23.0
google-cloud-logging==3.13.0
google-auth==2.48.0
python-dotenv==1.2.1
pydantic==2.10.0
pydantic-settings==2.7.0
pyjwt[crypto]==2.11.0
